+++ public/app/features/dashboard-scene/scene/layout-default/DefaultGridLayoutManager.tsx
@@ -28,6 +28,7 @@ import {
 } from '../../utils/utils';
 import { TabsLayoutManager } from '../layout-tabs/TabsLayoutManager';
 import { DashboardLayoutManager } from '../types/DashboardLayoutManager';
+import { LayoutRegistryItem } from '../types/LayoutRegistryItem';
 
 import { DashboardGridItem } from './DashboardGridItem';
 import { RowRepeaterBehavior } from './RowRepeaterBehavior';
@@ -45,7 +46,7 @@ export class DefaultGridLayoutManager
 
   public readonly isDashboardLayoutManager = true;
 
-  public static readonly descriptor = {
+  public static readonly descriptor: LayoutRegistryItem = {
     get name() {
       return t('dashboard.default-layout.name', 'Default grid');
     },
@@ -54,6 +55,7 @@ export class DefaultGridLayoutManager
     },
     id: 'default-grid',
     createFromLayout: DefaultGridLayoutManager.createFromLayout,
+    kind: 'GridLayout',
   };
 
   public readonly descriptor = DefaultGridLayoutManager.descriptor;
